Consolidate-the-duplicated-hooks
Consolidate-the-duplicated-hooks
