+++ frontend/README.md
@@ -29,6 +29,10 @@ To learn more about Next.js, take a look at the following resources:
 
 You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js) - your feedback and contributions are welcome!
 
+## Hooks Convention
+
+All custom hooks live in `frontend/lib/hooks/` and follow camelCase naming (e.g., `useLogin.ts`, `useAuthInit.ts`). Do not create a `frontend/hooks/` directory.
+
 ## Deploy on Vercel
 
 The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
